fix(dev): find correct document by slug (#498)
There were some recent changes in the Yext CLI generate-test-data
command. Beforehand we were blindly returning the first document
returned from GTD, which led to inconsistencies because the wrong
document could have been selected. Since the actual document we care
about cannot be determined on the CLI side, we need additional logic to
figure out which document to return based on the slug value and slug
field used based on the templates available.