Addressed comments.
Tested with npx pages dev --module <widget> using both folder name
and module.config.name matching.

Regardles of the method used, if a config is found, the appropriate
URL is opened and the module is loaded.

Ex: a module with name
"prefix" is under a folder named "tailwind". Running npx pages dev
--module prefix or npx pages dev --module tailwind both opens
localhost:5173/modules/prefix and the module loads as expected.

Tested with module that does not have postcss and bad module name.

When a bad module name is used, an error log "Could not find
module for "<module>"" is logged, and the server does not start.

npx pages dev works as expected.

If using "./src/modules" in tailwind.config.ts, modules load much
faster. Will create anothe PR to address the scaffolding command.