feat: support scope with modules (#502)
`npm run prod -- --scope myScope` works w/ myScope in templates,
modules, and both. npm run build and dev work as expected as well.
Double checked scaffolding supported scope still too.

scope means files in both myScope and src/... are built and used.
Without the scope being set, the files in myScope are not used. If a
file is in myScope and in src/... (with same config name), the one in
myScope should be used (if scope is set).

---------

Co-authored-by: Matt Kilpatrick <mkilpatrick@yext.com>