fix(dev): adjust GTD logic and pass scope (#504)
Fixes the following issues:
1. Not passing the --scope flag to the CLI GTD which is necessary for
config v2 (we were only passing --hostname which works the same but only
for config v1)
2. `slug` wasn't added to the --slugFields flag if no other alternate
slugFields were used
3. Filters out static templates so we can actually find the correct
document. Currently, the --featuresConfig flag is not honored by the CLI
so static template document data is coming back when it shouldn't. This
is a workaround for that.