fix(plugin): non-scoped templates not rendering properly (#517)
When rendering with a scope, non-scoped templates weren't getting their
css added to the head. This is because there was previously an
assumption that all templates were always in the scoped folder. We now
correctly look for the template path in the manifest before returning
which template path to use (scoped or not).