fix(modules): update scaffold tailwind  (#553)
Updating how scoping tailwind logic for modules work. Confirmed it works
in local dev, local prod, and platform. Tested with multiple modules on
one page as well ensuring that tailwind styles don't leak into inserted
page.

Instead of needing 

```
.tailwind {
   @tailwind base;
}
```

in the index.css we are using the `tailwindcss-scoped-preflight` library
in the `tailwind.config.ts ` to scope it to `tailwind tw-${moduleName}`.

So the module component will need to be wrapped by  

```
<div className="tailwind tw-${moduleName}">
  Module
</div>
```