feat(dev): support visual editor local dev mode (#572)
Adds support for in-platform page sets to pages js dev sever

- When the dev server starts with the `--siteId` flag passed, the
in-platform page sets are fetched for that site.
- When the index page is loaded, each up to 10 documents are fetched for
each page set and listed in a table.
- When `--prod-url` is true (default), requests to slugs first try to
look up the in-repo page. If that fails, then it attempts to fetch an
in-platform document with that slug from each page set until a match is
found.
- When `--no-prod-url` is true and when POST requests are made (occurs
for the VE preview), the url is first parsed as `pageSetId/entityId` and
it attempts to fetch an in-platform document for that combination. If it
is not found, then the normal in-repo lookup occurs.


This is what previously didn't work in dev mode:
1. Live preview
   1. No way to inspect the actual output HTML document
2. No entity pages are served at localhost:5173
    1. the in platform page set pages do no appear on the index page
3. Testing changes to main.tsx
   1. Slug paths, head config, other wrappers
4. Testing adding a page set or changing a page set’s scope 
    1. (without it going live in production)

1, 2, and 3 now work

See https://github.com/yext/visual-editor/pull/251



https://github.com/user-attachments/assets/f7157d21-0409-42c5-9c7e-dc31e95d4537

---------

Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>